requirements.txt
PySide6>=6.0.0
websocket-client>=1.6.0
requests>=2.28.0

pyinstaller>=6.0.0